Fix Grid Design + Added New Pricing Section
Was having an issue with the formatting of the testimonials grid. Was not being created in a row as code intended, but was resolved with cleaning up code and correcting spelling mistake.

I have also started to build out the pricing section, and creating HTML for the card and content.