FAQ Section
Started to add FAQ section to the website. Just need to style it.

Had an issue with the border bottom of each faq, as was incorrectly appearing at the bottom of the section rather than the bottom of each FAQ.

Resolved this by correcting the way each element was nested, as <div> was incorrectly nested and not closed.